# api/swagger-api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Поиск мероприятий",
        "description": "Возвращает список мероприятий с фильтрацией по ID, названию, категории, цене, городу, датам и организатору.",
        "operationId": "listEvents",
        "parameters": [
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "Поиск по точному ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "title",
            "in": "query",
            "description": "Поиск по подстроке названия",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Фильтр по категории (meetup, concert, exhibition, party, other)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "price_from",
            "in": "query",
            "description": "Минимальная цена",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "price_to",
            "in": "query",
            "description": "Максимальная цена (price_to=0 для бесплатных)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "city",
            "in": "query",
            "description": "Город проведения",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "description": "Дата начала не раньше (YYYYMMDD)",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": 20260314
          },
          {
            "name": "date_to",
            "in": "query",
            "description": "Дата начала не позже (YYYYMMDD)",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": 20260314
          },
          {
            "name": "user",
            "in": "query",
            "description": "Никнейм организатора",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Лимит пагинации",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 10
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Смещение пагинации",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 0
          },
          {
            "name": "X-Session-Id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Результаты поиска",
            "headers": {
              "Set-Cookie": {
                "description": "Возвращает куку БЕЗ обновления TTL",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "example": {
                  "message": "invalid \"limit\" parameter"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Создание события",
        "description": "Доступно только авторизованным пользователям. Создает новое мероприятие и привязывает его к текущему пользователю.",
        "operationId": "createEvent",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "cookie",
            "description": "ID сессии из Cookie",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "3f8a2c1d9e4b7f0a5c6d2e8b1a3f9c7d"
          }
        ],
        "requestBody": {
          "description": "Данные нового события",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreateRequest"
              },
              "example": {
                "title": "Мой день рождения",
                "address": "г. Санкт-Петербург, ул. Пушкина, дом Колотушкина",
                "started_at": "2026-04-01T12:00:00+03:00",
                "finished_at": "2026-04-01T23:00:00+03:00",
                "description": "Приглашаю вас отпраздновать мое 30-с-чем-то-летие"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Событие успешно создано",
            "headers": {
              "Set-Cookie": {
                "description": "Обновляет TTL сессии",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "example": {
                  "id": "12e9c0b1a2b3c3d5e6f7a8b7"
                }
              }
            }
          },
          "400": {
            "description": "Невалидные входные данные",
            "content": {
              "application/json": {
                "example": {
                  "message": "invalid \"title\" field"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "Событие с таким названием уже существует",
            "content": {
              "application/json": {
                "example": {
                  "message": "event already exists"
                }
              }
            }
          }
        }
      }
    },
    "/events/{id}/like": {
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Лайк на мероприятие",
        "description": "Позволяет авторизованному пользователю поставить лайк мероприятию.",
        "operationId": "likeEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID мероприятия",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "12e9c0b1a2b3c3d5e6f7a8b7"
          },
          {
            "name": "X-Session-Id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Лайк успешно поставлен",
            "headers": {
              "Set-Cookie": {
                "description": "Обновляет TTL сессии",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Мероприятие не найдено",
            "content": {
              "application/json": {
                "example": {
                  "message": "Event not found"
                }
              }
            }
          }
        }
      }
    },
    "/events/{id}/dislike": {
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Дизлайк на мероприятие",
        "description": "Позволяет авторизованному пользователю поставить дизлайк мероприятию. При этом лайк (если он был) удаляется.",
        "operationId": "dislikeEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID мероприятия",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "12e9c0b1a2b3c3d5e6f7a8b7"
          },
          {
            "name": "X-Session-Id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Дизлайк успешно поставлен",
            "headers": {
              "Set-Cookie": {
                "description": "Обновляет TTL сессии",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "headers": {
              "Set-Cookie": {
                "description": "Удаляет куку (Max-Age=0)",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Мероприятие не найдено",
            "content": {
              "application/json": {
                "example": {
                  "message": "Event not found"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventCreateRequest": {
        "type": "object",
        "description": "Запрос на создание нового события",
        "properties": {
          "title": {
            "type": "string",
            "description": "Название события",
            "example": "Мой день рождения"
          },
          "description": {
            "type": "string",
            "description": "Описание события",
            "example": "Приглашаю вас отпраздновать мое 30-с-чем-то-летие"
          },
          "address": {
            "type": "string",
            "description": "Адрес проведения",
            "example": "г. Санкт-Петербург, ул. Пушкина, дом Колотушкина"
          },
          "started_at": {
            "type": "string",
            "description": "Дата и время начала в формате RFC3339",
            "example": "2026-04-01T12:00:00+03:00"
          },
          "finished_at": {
            "type": "string",
            "description": "Дата и время окончания в формате RFC3339",
            "example": "2026-04-01T23:00:00+03:00"
          }
        },
        "required": [
          "address",
          "finished_at",
          "started_at",
          "title"
        ]
      },
...
    }
  }
}